feat: add auth to the backend adsfasdfasdfsafsdfas
In this commit we are adding auth endpoints for google
api to hit once a user signs in from the front end.
Specifically when the `login` endpoint is hit, it validates
the email to check if its allowed, then creates a Cookie
for the User to use in subsequent api calls.